// src/main/interfaces/responseIf.ts
export interface formatWinnerInfoIf {
  tableId: string,
  userId: string,
  seatIndex: number,
  winAmount: number,
  // balance: number
  isCash:boolean;
  coins:number;
  cash:number;
  winCash:number;
  bonus:number;
}